api/requirements.txt
autogluon.timeseries==1.2
fastapi==0.115.8
pandas==2.2.3
uvicorn==0.34.0